check account

+++ post-app/app/Http/Controllers/PostsController.php
@@ -124,6 +124,11 @@ class PostsController extends Controller
 
         $post = Post::find($id);
         // 이미지 파일 수정. 파일시스템에서
+        // Authorization. 즉 수정 권한이 있느지 검사
+        // 즉, 로그인한 사용자의 게시글의 작성자가 같은지 체크
+        if(auth()->user()->id != $post->user_id) {
+            abort(403);
+        }
         if($request->file('imageFile')) {
             $imagePath = 'public/images/'.$post->image;
             Storage::delete($imagePath);
@@ -145,6 +150,17 @@ class PostsController extends Controller
        // 파일 시스템에서 이미지 파일 삭제
        // 게시글을 데이버베이스에서 삭제
        $post = Post::findOrFail($id);
+
+        // Authorization. 즉 수정 권한이 있느지 검사
+        // 즉, 로그인한 사용자의 게시글의 작성자가 같은지 체크
+        // if(auth()->user()->id != $post->user_id) {
+        //     abort(403);
+        // }
+
+        if ($request->user()->cannot('update', $post)) {
+            abort(403);
+        }
+
        $page = $request->page;
        if ($post->image) {
            $imagePath = 'public/images/'.$post->image;
